Use more precision for transfer functions
Some standards such as ACEScct specify transfer function constants with
more precision than can be represented with floats.

+++ src/commonMain/kotlin/com/github/ajalt/colormath/internal/InternalMath.kt
@@ -33,4 +33,4 @@ internal inline fun <T> fromPolarModel(c: Float, h: Float, block: (a: Float, b: 
 /**
  * return `sign(a) * |a|^p`, which avoids NaN when `this` is negative
  */
-internal fun Float.spow(p: Float) = absoluteValue.pow(p).withSign(this)
+internal fun Double.spow(p: Double) = absoluteValue.pow(p).withSign(this)
